pyproject.toml
[project]
name = "rag-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "boto3>=1.34.0",
    "tqdm>=4.66.0",
    "openai>=1.101.0",
    "python-docx>=1.1.0",
    "chromadb>=0.5.5",
]

[project.scripts]
textract-ocr = "textract_ocr.cli:main"
llm-process-text = "textproc.cli:main"
split-text-files = "textsplit.cli:main"
merge-split-files = "textsplit.merge_cli:main"
docx-to-text = "docx_text.cli:main"
legal-chunk-text = "textsplit.legal_cli:main"
vector-db = "vectordb.cli:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/textract_ocr", "src/textproc", "src/textsplit", "src/docx_text", "src/vectordb"]

[project.optional-dependencies]
clean = [
    "openai>=1.101.0",
]
